Deleting files now is accurate
Bug existed that looked at any file beginning with user UID, then
deleted it.  So if user UID was 10, delete would delete 10, 100, 101,
102 etc.  Now, checks are made to ensure only files beginning with
‘UID_’ are deleted.  Fixes #44

+++ modules/students/actions/delete.php
@@ -1,33 +1,40 @@
 <?php
 require_once($_SERVER['DOCUMENT_ROOT'] . '/engine/initialise.php');
-$user = Students::find_by_uid($_POST['uid']);
 
-if ($directory = opendir(SITE_LOCATION . "/uploads/")) {
-	$uidLen = strlen($user->uid);
-	$userPhotoPrefix = $user->uid . "_";
+if (isset($_POST['uid']) && $_POST['uid'] <> '') {
+	$user = Students::find_by_uid($_POST['uid']);
+	printArray($user);
 	
-	while (false !== ($entry = readdir($directory))) {
-		$photoUserID = substr($entry, 0, $uidLen) . "_";
+	if ($directory = opendir(SITE_LOCATION . "/uploads/")) {
+		$uidLen = strlen($user->uid);
+		$userPhotoPrefix = $user->uid . "_";
 		
-		if ($photoUserID == $userPhotoPrefix) {
-			unlink(SITE_LOCATION . "/uploads/". $entry);
+		while (false !== ($entry = readdir($directory))) {
+			$underScorePos = strpos($entry, "_");
+			$photoUserID = substr($entry, 0, $underScorePos) . "_";
 			
-			$log = new Logs();
-			$log->type = "info";
-			$log->title = "File Delete";
-			$log->description = "File " . $entry . " was deleted";
-			$log->create();
+			if ($photoUserID == $userPhotoPrefix) {
+				unlink(SITE_LOCATION . "/uploads/". $entry);
+				
+				$log = new Logs();
+				$log->type = "info";
+				$log->title = "File Delete";
+				$log->description = "File " . $entry . " was deleted";
+				$log->create();
+			}
 		}
+		
+		closedir($directory);
 	}
 	
-	closedir($directory);
+	$user->delete();
+	
+	$log = new Logs();
+	$log->type = "warning";
+	$log->title = "User Deleted";
+	$log->description = "User UCAS:" . $user->ucas . " was deleted";
+	$log->create();
+} else {
+	echo "No UID provided!";
 }
-
-$user->delete();
-
-$log = new Logs();
-$log->type = "warning";
-$log->title = "User Deleted";
-$log->description = "User UCAS:" . $user->ucas . " was deleted";
-$log->create();
 ?>
